+++ RoundParser/src/main/java/com/marco/csgoutil/roundparser/repositories/interfaces/RepoUserScore.java
@@ -1,5 +1,6 @@
 package com.marco.csgoutil.roundparser.repositories.interfaces;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.marco.csgoutil.roundparser.model.entities.DaoGames;
@@ -50,5 +51,5 @@ public interface RepoUserScore {
 	 * @param steamID
 	 * @return
 	 */
-	public List<Double> getLastXUserScoresValue(Integer counter, String steamID);
+	public List<BigDecimal> getLastXUserScoresValue(Integer counter, String steamID);
 }
